One-to-many support for JDO
Sadly, getting 1-to-many Lists working for JPA did not get me nearly as
close to having 1-to-many Lists working for JDO.  This CL takes it the
rest of the way.

I'm leaving one test failing pending a DataNucleus change from Andy.
I'll fix it up once I've synced the ORM to DataNucleus head.