Support unencoded String and Long primary keys
Massive change based on discussions with kgibbs and ryanb.

Introduces support for Long pks and unencoded String pks.  Only legal
for classes that do not have ancestors.

Unencoded String pks is now the default for a String primary key.  If
you want the old behavior where the String is an encoded representation
of a Key, annotate the field with "encoded-pk".

Also changes the way you provide an ancestor when using a Key primary
key.  If you want to specify an ancestor and get id generation, you use
the ancestor-pk field.  If you want to specify an ancestor and provide a
named key, you use the key field directly.